IJPL-173942 [reworked terminal] Fix line length detection
The length of the editor line for the purpose of calculating
soft wraps may or may not include the scrollbar width.
The actual logic is quite complicated, and there's a workaround
in the editor that aimed to prevent soft wraps from getting
under the scrollbar.

However, in the terminal, we actually do want soft wraps
under the scroll bar, because otherwise the effective width
would change depending on whether the scroll bar is visible.
This would lead to width changing when entering / leaving
the alternate buffer mode, which causes various glitches.

Fix by introducing an internal API to allow soft wraps under the scroll bar.

GitOrigin-RevId: b2b8a219dceac919dc6ebd40963b0abab97937a7