IJPL-175730 Fix Access is allowed from EDT only thrown from `InfoAndProgressPanel.addProgress`
In afd2fdb2be87f896fe0a14ba99ff6e90365ba58e I've moved `registerCloneTasks` from IdeStatusBarImpl constructor to a separate `initialize` method which is called from `ProjectFrameHelper.setProject`.

`setProject` is not invoked on EDT, which leads to the exception being thrown when the "clone" progress is trying to be added to the status bar.
The exception only happens with `rhizome.progress` registry flag turned off.

To fix that, I moved `statusBar.initialize` to `setRawProject` under `withContext(Dispatchers.EDT)`
`setRawProject` is actually a better place to call initialize as it's guaranteed that `windowManager.assignFrame` is going to be called before that.

It's important because we need this method to be called in order to properly display all progresses

GitOrigin-RevId: fca7df764e7f267f843292efbea4d752e3b96577