IJPL-176002 [reworked terminal] Fix SIOOBE in TerminalCursorPositionTracker
Due to async updates of the text buffer contents and cursor position,
it's possible that the cursor position can be outside the current line.
To avoid throwing exceptions, don't look beyond the end of the line
for the purpose of counting double-width characters.

Note that there's another similar bug when the line doesn't
reflect the actual terminal line, being cut off one character
short on Windows. It leads to a similar exception,
but it's not fully fixed here. This fix will prevent the backend
from getting into a broken state, but the frontend will then
display the cursor in the wrong position or throw an exception.
This issue should be fixed separately.

GitOrigin-RevId: 145d52b66a5a0f273e9cc1d33eb0d1f10208093d