[terminal] IJPL-157242 Start terminal session using new protocol if Reworked Terminal is enabled
Unfortunately, the code that creates the TerminalWidget and starts the session can't be fully extracted to the frontend part because there are a lot of internal and external usages (`TerminalToolWindowManager`).
So, I had to create a hacky `TerminalWidgetProvider` extension point to get the `ReworkedTerminalWidget` from the frontend.
`TerminalToolWindowManager` should be mostly called on the frontend, but some calls can occur on the backend. For now, there is no logic to catch these cases, but probably we need it. Need to consider later.

GitOrigin-RevId: b3c528921442268e81f3f6fb094b66ad73720f94