IJPL-165667 [reworked terminal] Make sense of the ClearBuffer action
Write a detailed comment on that 0-vs-1-based indices issue.

Set the cursor to 1, as setting it to 0 made little sense and only
worked because the implementation would update it to 1 anyway,
but that's fragile.

Add a new line first, though, so we're putting the cursor on the line
that already exists. Otherwise, the cursor would create an empty
line by itself first and then the line we add would become the second one.

GitOrigin-RevId: 06867c29ea6348f28d2f11fa6a6f3f089f6493eb