DBE-22700: Prepare Grid modules for platformization
grid.impl depended on platform.images. Platform module can't depend on a plugin, and it's risky to make platform.images a platform module now. So decision is to separate the part of grid.impl that depended on platform.images to a dedicated module which will NOT be a platform one.

Also, intellij.grid API module was converted to V2 that will simplify platformization

GitOrigin-RevId: 3dea3dd0b5e05a283ee3c26b5806fa33822b3e8a