IJPL-173817 [reworked terminal] First search implementation
The session was copied and adapted from the gen1 terminal.

The controller had to be created separately, as we need something
to put that logic into, and we don't have anything like that
in the reworked terminal yet. Also, we need an entry point
to put into the data context.

The CloseSearch escape handler is implemented.

The TerminalPanel class had to be converted into
a JBLayeredPane to reuse the UI design from gen1.
Laying out the editor here is much simpler, though,
as it simply takes the entire space.

GitOrigin-RevId: b6dcc8b0562ca3c0527eaf329755f22265f82cb9