[java] cleanup: get rid of fake packages in content modules of Java Frontend plugin (IDEA-366600)
Fake packages were used to ensure that the main plugin classloader is used to load the classes from content modules. Now the same can be achieved explicitly with loading="embedded" attribute. This is also needed to ensure that the classpath for content modules can be properly specified (this is needed for IJPL-165692).

Also, it's possible to use auto-layout for the plugin, so there is no need to specify it explicitly.

GitOrigin-RevId: 6a75b4868b796215018d230949752bc978c60476