[terminal] IJPL-176339 Add spaces to the document line if the cursor position is out of its bounds
JediTerm implementation can move the cursor to any place of the screen, even if there is no content.
We already have a guarantee that line with cursor exists in the document, but we don't have a guarantee for the column.
So, we can receive cursor position change event with column value out of line bounds.
Let's try to add spaces to the document till the cursor position and see how it goes.

GitOrigin-RevId: 08e1a43366795c0a315825b6de3a782469f3ab30