[yaml] properly split YAML plugin and its frontend variant by modules (IJPL-176102)
All code from the YAML and 'YAML Extension for Remote Client' plugins is moved to content modules. This is needed to avoid duplication in plugin.xml files, enable automatic layout, and to allow merging these plugins to a single one.

'intellij.yaml' is registered as an embedded content module, so it uses the main plugin classloader, and now has a dependency on 'intellij.yaml.editing' module. This ensures that classes from both modules are available to content modules from other plugins which declare dependency on YAML plugin as before.

GitOrigin-RevId: c6d2a156d2ff60aca78925e2e3ef7b96f6593e5a