IJPL-84874 Make SmartExpander respect ide.tree.collapse.recursively
This turned out harder than it would seem, as SmartExpander
can't access Advanced Settings due to module dependencies.
So we expose it via an ugly internal API and then set it
in the listener when the setting changes and in the tree
constructor to set the initial value.
The tree constructor doesn't seem the right place to do it,
but where else? We need this code to be invoked before
the user starts expanding / collapsing trees.

GitOrigin-RevId: 5cb7b023c6817969b31ba6c73b79b487d17653db