IJPL-175562 Cell grid text rendering: simple fragments, first attempt
Grid calculation is based on character widths here.

First, if the character width is almost-divisible by the cell width,
no adjustment is performed. This is a very common case
with monospaced fonts, so it's a good optimization.

Then, the number of slots the character takes is calculated
based on its width and cell width, dividing these and rounding up,
but with a 10% threshold. This threshold allows skipping
slightly wider characters just in case.

When painting, characters without alignment are painted
as-is, while the aligned characters are painted one-by-one,
centering them in their cells.

GitOrigin-RevId: 4e2f3a46eaa8c822e45639f577988c0cad414716