IJPL-175562 Cell grid text rendering: complex fragments, first attempt
Grid calculation is based on glyph positions here.
It is assumed that one glyph is one character.
This will likely have to be improved in the future.

Unlike simple fragments, we start with absolute positions here,
not widths. Therefore, when we perform adjustments,
we need to take care to shift the following glyphs accordingly,
even if those glyphs are not affected by alignment.

GitOrigin-RevId: 8460c1143858235100bb2640ca881b2b39427db2