[terminal] IJPL-176464 Scroll terminal down till the last non-blank line or cursor
Previously, we took into account only the cursor position, but there might be a case when some meaningful output appears below the cursor position. For example, completion output in Zsh.
So, I had to adjust the logic to scroll to the last non-blank line or cursor position, depending on what is lower.
Also, I had to add the listener that updates the scroll position on document updates, because in case of completion cursor doesn't move, so we don't update the scroll.

GitOrigin-RevId: f33682a253bf42cddfe7de25e6cb459422733267