[build scripts] add JARs from lib/modules/ subdirectory of a plugin directory to classpath when running tests
JAR for a content module can be placed under lib/modules/ directory if the package prefix isn't specified for it. So we need to add such JARs to the list returned from 'createIdeClassPath', otherwise Rider tests fail if some modules in YAML plugins are extracted to separate JARs (this is needed for IJPL-176102).

GitOrigin-RevId: 22f0f57a6c927a6ca261b849dc3428851bbaef07