IDEA-366795: Set context when creating new declaration
When running in Android Studio against a non-physical type (eg, a
data-binding or `R` class) that uses a scope enlarger, this inspection
is producing a new variable declaration that uses a fully-qualified type
name when in fact it should be a shortened name.

Creating the declaration with an appropriate context element will allow
the correct form of the type name to be generated.