IJPL-84874 Introduce TreeState API for flattened nodes
Use case: nodes like "a/b/c" that correspond to nested
"a" / "b" / "c". If the representation is different when the state
is stored/restored, then a simple one-to-one approach doesn't work.

The new API allows nodes explicitly report their unflattened
values, so they're stored instead of the flattened node.

The API also extended to support implementations of the
PathElementIdProvider interface by nodes instead of user objects.
This is needed for trees that don't use node descriptors
and instead follow the classic Swing approach when the node
is a DefaultMutableTreeNode and the user object is something
that has nothing to do with UI whatsoever. A good example
is VCS trees, for which this API is being introduced now.

GitOrigin-RevId: 9082a6a84294efee0540444c4e5b4e08f00ee5bc