IJPL-84874 Use the new PathElementIdProvider API for VCS changes
With this API it's now possible to maintain the expanded state
even if a directory is flattened / unflattened.

Note that it still doesn't work when a file is moved from
the staged to the unstaged tree or vice versa.
In that case the paths are different, so moving a file
from an expanded directory often leads to the same directory
being collapsed in the new location.
This is a separate issue even though it may look similar.

GitOrigin-RevId: 32b831e5708ee374d15a62d0e43f7a22502f5d8b