PY-54269 Remove two confusing tests in PySdkPathsTest replicating the internals of adding user interpreter paths
It seems that the original problem is no longer relevant now that
editable and regular SDKs don't share additional data.

See 126342bee2fdb1ec38c8b03623450dd70f847b7b.

The tests relied on the fact that the previous logic of setting
"transferred paths" might have stored them in additional data
of an "editable" copy of an SDK not actually specified as an SDK
of any module (hence SDK additional data was updated but not actual
project roots). See how PyTransferredSdkRootsKt.setPathsToTransfer
is called in `PythonSdkUpdater.updateSdkPaths`.

The new mechanism that also considers `sys.path` entries for
editable cross-module dependencies doesn't have this side effect,
and breaks the tests.

Namely, depending on whether `ProjectJdkTable.updateJdk(sdk, editableSdk)`
or `ProjectSdksModel.apply()` is called between the introspection
of the editable and the actual SDK, copying the additional data of
the first to the second, existing "tranferred roots" might get out of
sync with actually configured source roots for the given SDK,
affecting setting and removing them on `sys.path` change.

GitOrigin-RevId: 3e3345d056f0a2671942bc121188d650b3e4c99e