[build scripts] exclude content modules from distribution if they aren't compatible with the current product mode (IJPL-165692)
ContentModuleFilter interface is introduced which can be used to filter out some optional content modules from the distribution. It's used to exclude modules not compatible with the frontend mode when building a standalone frontend distribution (JetBrains Client). This is needed to be able to use the same plugin.xml for the backend and the frontend parts without increasing the size of the standalone frontend distribution.

ContentModuleFilter can be also used for other cases when we need to exclude unavailable modules from the distribution, see IJPL-176806.

GitOrigin-RevId: 9e6b6dac1753d586199c39e96998760df17db5c7