[build scripts] skip unresolved optional content modules when building IDEA Community from intellij-community source code (IJPL-176806)
We have cases when a variant of IDEA Community should be built from a subset of ultimate sources, and some content modules are missing in that subset. Also, this is needed to use the same plugin.xml for the frontend and the backend variant of a plugin (IJPL-165692) if the plugin is located in intellij-community (e.g., YAML), because the frontend variant uses API which is currently not available in intellij-community.

Here a new property 'skipUnresolvedContentModules' is added to ProductModulesLayout to support such cases.

GitOrigin-RevId: 235fa3e05078ff95150dc0180634adc2c383388d