IJPL-562 StubElementFactory#shouldCreateStub: make `node` parameter not-null
There are a couple of calls that potentially pass null as an argument.
But the `null` argument does not make sense, the result is always `false` in that case.
So, I'm making the parameter notnull and updating the calls that potentially pass null.

(cherry picked from commit 0b36a4d5838a65424ecd9f5a34a0e4d6ea1ea196)

# Conflicts:
#	community/platform/core-impl/api-dump-experimental.txt
#	phpstorm/php/src/com/jetbrains/php/lang/inspections/reference/PhpRefManagerImpl.java
#	phpstorm/php/src/com/jetbrains/php/lang/psi/stubs/stub_factories/PhpDocCommentStubFactory.kt
#	phpstorm/php/src/com/jetbrains/php/lang/psi/stubs/stub_factories/PhpDocTagStubFactory.kt
#	phpstorm/php/src/com/jetbrains/php/lang/psi/stubs/stub_factories/PhpUseListStubFactory.kt
#	phpstorm/php/src/com/jetbrains/php/lang/psi/stubs/stub_factories/PhpUseStubFactory.kt
#	phpstorm/php/src/com/jetbrains/php/lang/psi/stubs/stub_factories/PhpVariableStubFactory.kt

GitOrigin-RevId: 2eab40b5dcf6134b68ad50d0fbab983476d337ec