IJPL-562 StubElementFactory#shouldCreateStub: make `node` parameter not-null
There are a couple of calls that potentially pass null as an argument.
But the `null` argument does not make sense, the result is always `false` in that case.
So, I'm making the parameter notnull and updating the calls that potentially pass null.

(cherry picked from commit 0b36a4d5838a65424ecd9f5a34a0e4d6ea1ea196)

GitOrigin-RevId: 910cf9ff009a72a10bceaa376d6c696736868828