IJPL-175562 Fix ComplexTextFragment glyph aligning
The previous implementation was bugged because
it was hard to keep track of what has changed and what hasn't.
Instead of trying to fix bugs in complicated code,
let's split it into three passes.

During the first pass, we just calculate the original widths,
nothing more. Very simple.

The second pass contains most of the logic that the previous
implementation had, but without centering, because centering
introduces those hard-to-reason-about position changes.
To simplify it further, instead of trying to calculate new positions
based on the old ones, calculate the new ones by summing up
widths.

The third pass is the centering one. Also very simple
to reason about after all the coordinates have already been updated.

GitOrigin-RevId: e2227931495baf66277b429a22d8a5c130429cb0