IJPL-175562 Account for alignment when calculating character positions
In ComplexTextFragment, when calculating character positions,
it's now necessary to account for glyph shifts caused by centering
characters within their cells. The visual end of a character is no longer
its rightmost position (for left-to-right), but often slightly further than
that because there might be empty space between glyphs,
but we want selection to extend to the end of the cell nevertheless.

GitOrigin-RevId: d13e5805ae2fbc545f777d1d25567250054096dd