IJPL-175562 Only perform char alignment if its slot is wider
The case when the slot is narrower is tricky and shouldn't really
occur in practise except with really strange characters or fonts.
The rest of the code doesn't work well with negative alignments,
so let's start by prohibiting them altogether and then fix issues
on a case-by-case basis if they appear.

GitOrigin-RevId: 67ed1f12a919946e131ba815b49655e0a7923af2