[yaml] get rid of an additional "YAML Extension for Remote Client" plugin (IJPL-176102)
'intellij.yaml.frontend.split' module is now registered in the main YAML plugin, and the plugin is reused for the frontend variant of the IDE.
Dependency on JSON is temporarily removed from the plugin.xml, because the corresponding module isn't available in the frontend IDE. It'll be replaced by a dependency on a shared JSON module later.
Since currently it isn't possible to specify that a content module is required when running in backend mode and isn't required when running in the frontend mode (see IJPL-177060), additional modules are registered as optional for now.

GitOrigin-RevId: e47ad87e2de4d34e5fa1b235d50c88a80cf949dc