IJPL-177074 Fix PopupShowOptions.aboveComponent for WizardPopup
When this new API was introduced, it was only used in
com.intellij.vcsUtil. However, it turns out that this package
is not used in just VCS, but in other places like AI Assistant.
And in some of those places not AbstractPopup, but WizardPopup
and its subclasses are used (like ListPopupImpl
and ActionGroupPopup). So the new API must also work for those.

Fix by only keeping the existing override logic in showImpl
for the case when the top-left anchor is used. This is the default,
so everything except aboveComponent is not affected.

GitOrigin-RevId: e20d1d8c5a3d7762833440477df61e9f2a4266f1