[platform tests] don't report unresolved optional content modules when checking intellij-community project (IJPL-176806)
We have cases when a variant of IDEA Community should be built from a subset of ultimate sources, and some content modules are missing in that subset. Also, this is needed to use the same plugin.xml for the frontend and the backend variant of a plugin (IJPL-165692) if the plugin is located in intellij-community (e.g., YAML), because the frontend variant uses API which is currently not available in intellij-community.

GitOrigin-RevId: a8b8e609bf4c3f2e95be4287e12a719915fe98a2