[rdct] IJPL-169760: Adds focus tweak for StructureView
Since a focused element is not generally available in the Remote Development, or it is set incorrectly (e.g., to the whole IDE itself), several fallbacks are used.

The first fallback is to set the focus owner component to the editor. However, this component does not always have a virtual file in the `DataContext `, so another fallback is added.

Additionally, a client id has been propagated to get the correct editors.

GitOrigin-RevId: 031a4cccf58d34525455eaa42fc9bd3dcf866a73