[recent files] IJPL-177136 Extracted and refactored selection preservation logic, which is now encapsulated in the custom SelectionModel implementation
Previously, a single-threaded sequential code was responsible for both populating data models and setting up focused and selected items in the UI lists. With async files fetching introduction, it is no longer possible to wait for all elements to be added into list before updating selection. This commit tries to introduce a selection model that ensures old UI behaviour while being capable of working with an asynchronously populated list model.

Its aims are:
- to disallow extending selection when elements are being added to the collection model one by one with delays in between
- to ensure selection existence after a selected element removal for smooth traversal experience

GitOrigin-RevId: 14f29dcf52eb2c541c5b5227e58808d8c5eb43b6