[terminal] Fix PowerShell virtual env customizer test
The meaning of parameters of `LocalShellIntegrationInjector.injectShellIntegration` was changed in the scope of IJPL-176934.
So, the test was running with the shell integration of the New Terminal (Gen1).
Update the values to run it with the Classic Terminal shell integration.
But maybe it is worth adding the same test that will run with Reworked Terminal (Gen2) shell integration.

GitOrigin-RevId: 0bd43d4e1f65c7e42600d18d3b99b4177ea2639d