[yaml] restore dependency on JSON plugin from the main module of YAML plugin (IJPL-176102)
Before, only 'intellij.yaml.backend' had such a dependency, and that module is registered as optional, so if the JSON plugin is absent, no errors were shown for the YAML plugin, but its crucial functionality didn't work. Probably, users weren't able to get in such a situation because the JSON plugin is marked as the essential plugin in most of the IDEs, but it's still better to restore such a dependency.

GitOrigin-RevId: f1373304e10fda4c48bb03774ba674b5e521956c