IJPL-177051 Use an explicit strategy to determine DWC
Guessing double-width characters by their real width
seems to not work for unusual characters even with fixed-width fonts.
The classic terminal had its own utility function to figure out
DWCs. It seems to be non-trivial, so let's introduce an internal API
to delegate to it instead of copy-pasting it into the editor.

GitOrigin-RevId: 4fb00bc83db8145d4019fc615a118061bac4e744