IJPL-177051 Fix calculating length for soft wraps in grid mode
EditorView.getCodePointWidth calculates the real width of
a character. However, in grid mode we force every character
into a cell, and therefore we should account for this
when calculating soft wraps. Otherwise, a line may be wrapped
even if there's space available or, conversely, extend beyond the view.

To fix this, move the width calculation to EditorView
and use the DWC strategy to determine how many cells
the code point needs and then use the cell width to calculate
the real width.

To avoid having too many null checks, make the strategy
not nullable and use a naive default implementation
in case nothing was set. Currently this implementation
is not used because only the terminal uses the grid mode
and it sets its own strategy.

GitOrigin-RevId: ecbe357b305fddf97671eafe65b91458d13d5d3a