[terminal] IJPL-157242 Close tabs that has no started terminal session
Previous approach closed the terminal session connected to the terminal tab. But it works only if session is already started for the tab. If we have a tab (restored on startup) and user didn't open it, the session is not started for it. User can close it and we should also remove it on backend.
So, create `closeTerminalTab` API call, that will terminate the session and remove the tab if session is started, or just remove the tab if it is not.

GitOrigin-RevId: a550c3b1c2513c9faf260b94fd481bd500904366