[terminal] IJPL-157242 Refactor: create tab first and then start the session for it
It simplifies the API by separating the meaning of `createNewTerminalTab` and `startTerminalSessionForTab`.
Also, it protects us from the case when terminal tab is created in the UI and closed before the session starts, because now we will create backend tab for it immediately after terminal widget creation instead of waiting for it to show on the screen.

GitOrigin-RevId: 96229065a50b25dfa1824d361a75a3dd3b8ee6d0