IJPL-176818 [reworked terminal] Update cursor on document changes
This fixes the case when the cursor offset hasn't changed because
something was appended to the document and the same number
of characters was removed from the beginning to maintain the size.
The offset stays the same, and therefore isn't emitted from the flow.
In such cases the cursor would only be repainted later due to
some other event, such as blinking.

GitOrigin-RevId: 5eb1dc1c6adfd3854c1290cd31806f48a1a0edee