IJPL-177527 Fix initial tool window layout order
When a project is opened the first time, it gets
the currently active tool window layout.
However, this layout doesn't always have
the correct tool windows order.
Moreover, as new tool windows are registered
(not always immediately), they get the maximum
order, which means some of them will have the same order.
This leads to an inconsistent layout, which breaks DnD.

Fix by normalizing the order when loading a tool window layout,
similarly to how it's now done for the layout saved in the workspace
(which is why this issue only affects the first project open).
Also always allocate a new order value to new tool windows,
that is, the maximum value plus one.

GitOrigin-RevId: 20548f01e165135456569d2df3fd0d047a0389a6