IJPL-177571 [reworked terminal] Copy on selection
Implement a handler for editors.
A handler is only active when the corresponding setting
is enabled and when the editor is focused.

On Linux, the selection is copied into the middle-click
clipboard. On other systems, the regular CopyPasteManager
is used. It may have an undesirable effect of spamming
the copy history with selection, but it allows for better
integration with the platform. This is a separate issue.

GitOrigin-RevId: 8fcf7b18229b51d134316b1317d7bfb15f22299f