IJPL-176358 [reworked terminal] Cursor correctness on document update
The document is updated, the cursor offset may be updated
in the same event as well, and then the listener is invoked.
But the cursor offset update is scheduled through the EDT
dispatcher, which may or may not invoke the collectors
immediately depending on the coroutine context
(modality, read/write actions etc.).
This leads to the curCursorState value being outdated
at the moment the listener is invoked, even though
the flow already has the correct value.

The proper fix would be to redesign the API in such
a way that the cursor updates are always emitted
if the effective cursor position changes.

For now, fix by explicitly updating the cursor offset
value on document changed listener.

GitOrigin-RevId: 273425c422089223e03106e520db863f8dc24917