[platform] IJPL-173816 support running filters when trailing text is removed from the document
The problem here was that removing trailing text changes `RangeMarker` (`com.intellij.execution.impl.AsyncFilterRunner.DeltaTracker.endMarker`) which was designed to track deletion from the document top. As a result, incorrectly calculated link ranges were applied to the editor.

This change introduces the ` trackDocumentChangesManually ` flag which replaces `RangeMarker` with manual tracking of the document.

GitOrigin-RevId: edd7376333c030fa20406bdbdf0a7953be4cf619