IJPL-164669 Make tool window stripe invisible if there are no buttons
When there are no buttons on a stripe, it becomes 1px wide.
This can be very annoying if it's right next to a vertical scrollbar
in a maximized window, as it makes targeting the scrollbar
with the mouse much harder.

We could fix this by ensuring that the stripe is 0px wide in this
case, but it would be a more fragile solution, as any changes
in the stripe layout can break it later.

The more robust solution implemented here is to hide
the stripe if it has no buttons. To do this, we introduce several
listeners to monitor the tool window buttons and the visibility
of the more button. Then this information is used to compute
the stipe visibility, along with other variables used for that
previously (the UI setting and overlay flag).

GitOrigin-RevId: 691e1b1c44c1d65f2437759d126968020f0a589e