Performance tests: cleanup performance snapshots from previous tests
A given performance test can do several attempts (BenchmarkTestInfo#attempts).
Every attempt is captured with IJ CPU profiler, and the result is stored in MEASURE123.jfr of WARMUP123.jfr.
We need to clean up all these snapshots when starting a new test, otherwise the snapshots from prev runs with a big index will be shown as artifacts of the consequent runs with fewer number of attempts.

GitOrigin-RevId: 466a8c272c2b0c8e29aaf18b7f63ddf2c61ae783