[kotlin] add workaround for `PSI and index do not match` in `K2ScriptingKaModuleFactory`
We cannot be 100% sure that a file is a script solely based on its extension.
Details explaining why are written in the comments of KTIJ-32922.
However, this is a workaround to minimize the consequences of KTIJ-32912, which should work in most cases.

This way, we will not call `ktFile.isScript()` for regular `.kt` files,
minimizing stub access, which was the cause of KTIJ-32912.

KTIJ-32912

GitOrigin-RevId: eb828715648a487bab7fd76e271641df43cc1bd2