IJPL-178348 [reworked terminal] Enable search in the alternate buffer
After the previous refactoring, it's almost a trivial thing to do.

Enable the action handler for the alternate buffer model.

Ensure that the previous session is closed on any editor
change, regardless of whether we're leaving or entering
the alternate buffer mode.

A side note: the session is now finished before
requesting focus in the editor. This isn't really necessary,
just makes the code easier to read and reason about:
we first remove the editor, close its session and only then
focus the new editor.

GitOrigin-RevId: 927769b3824fafb8f9135e6dcc56e6cde2c53f7d