IJPL-178323 [reworked terminal] Fix the editor's Find action
Well, it's a tricky one.

We have our own TerminalFindAction to enable keymap customization.
It works just fine, and the corresponding editor's action shortcut
is blocked with our key event dispatcher.

However, the action could be invoked in other ways,
from the menu or Find Action. And then the "normal" editor
action is invoked because we don't override the handler.

To fix, override the handler and at the same time keep our own
action to enable shortcut customization. This way only our
action can be invoked via a shortcut, but when the editor
action is invoked some other way, it still delegates to our implementation.

GitOrigin-RevId: 21dbe92785b90a2e3889384c6986c1b80f58071a