[IFT] IJOB-586 Do not reference Recent Files frontend classes in lessons
Features Trainer plugin is not split to frontend and backend. So, it is acting fully on backend and can't access frontend classes (Recent Files feature was recently split). Since we don't have plans to adapt the Features Trainer to RemDev, let's do not reference the frontend classes to make it work without errors at least in monolith mode.

GitOrigin-RevId: f0aaaf97744fcd71c8612f817661f7de925f750c