IJPL-178555 ProjectWindowCustomizerIconCache: fix subscription
LafManagerListener.TOPIC is only broadcast on the app bus,
so the listener was never invoked on a LaF change,
and the cache was never cleared.

While we're at it, replace simple connections with properly
scoped ones. It's not strictly necessary for the project connection,
but it is necessary for the new app connection to avoid leaks,
so let's handle both the same way for consistency.

GitOrigin-RevId: c58d689f94c91eb8951ce4215daf3f80dfe18483