IJPL-168110 Delegate navigation to a directory to `PsiNavigationSupport`
In RD we have a custom implementation of this service which handles navigation to directories in the project view.
Delegating the navigation to the service makes package navigation work in the RD scenario.

We also should select a proper pane id to update (instead of always falling back to `ProjectViewPane.ID`)
It'll make the navigation work properly in CWM as well

GitOrigin-RevId: cf177ab3ebcc1c21da018b4630339750564f3989