KTIJ-31353 [kotlin] Handle odd PSI elements on `UnresolvedReference` diagnostics for complex assignments
This is currently the most straightforward approach
to handling this on the IDE side; in the future,
the PSI in the diagnostic can be adjusted to be
more aligned with other diagnostics reported
on the operator calls.

See KT-75331 for the details.

^KTIJ-31353 Fixed


(cherry picked from commit 338855f17f5dfafafe7cb5095455675e0f83940f)

IJ-CR-155970

GitOrigin-RevId: 8be4868860d16c9214f929d72d81ab564b02ca45