IJPL-178629 Fix index in ComplexTextFragment
The glyphs are indexed from 0 to count minus one,
the chars are indexed starting with the start index.
This was not taken into account, and therefore the logic
only works when start == 0 or, which is more often the case,
when the character at the wrong index has the same width
as the right one.


(cherry picked from commit 8c45828e6289ecf43db44804734526d5e4869c55)

IJ-CR-155687

GitOrigin-RevId: aa0c240f1c79507bd8e2f9052487bf73e8ae885c