[tabs] IJPL-178708 Fix tabs bouncing when 'Show pinned tabs in a separate row' option is enabled
All scroll changes were performed in a `scrollToSelectedTab` function.
If tab left part is located on the left of the visible tabs area, we try to adjust the scroll position. The same for the right tab part - we try to adjust the scroll to place it in the bounds of the tabs pane.
But, the code that tries to adjust the right tab position was triggered even if it is already placed at x = minX. So, we moved it to the left, then other `if` branch moved it to the right.
The solution here is to not try to move it when scroll offset is equal to minX position.

GitOrigin-RevId: f5ff305ea51772bdbc4b7fa73d57310d2f405a3f