IJPL-178555 Remove Graphics translation from AbstractToolbarComboUI
The problem with Graphics translation is that once there's
a DeferredIconImpl among the icons, it can use the coordinates
passed to paintIcon to schedule a repaint request. If these coordinates
are not relative to the component's coordinate system,
then the request will be wrong and the next repaint may be
clipped incorrectly.

Until this issue is fixed, we can at least temporarily fix
the issue with these toolbar combos by removing the translation
and basing the paint rectangle on the component's coordinate system.

GitOrigin-RevId: 124b5656b2101ac2e235a2d658c793a20d9208f6