IJPL-178858 Adjust main toolbar icons for both dark and light headers
The origin of the issue is not clear, but for some reason dark icons
get stuck in the dark state when the IDE was started in a dark theme.
Because IconLoader has both the dark and the light replacers,
though, always using it regardless of the current theme seems
to fix the issue.

There should be no performance issues because these icons
are cached anyway.

GitOrigin-RevId: 152a21c1793e8c92c25b0a22d2aebea62840bd29