Revert enforcement of AboutDialog EP contract
Turns out that the lack of enforcement and documentation of the
contract of this EP is relied upon by the Kotlin plugin, so we
need to avoid enforcing it. The EP KDoc has also been updated to
match the actual behaviour.

closes https://github.com/JetBrains/intellij-community/pull/2964


(cherry picked from commit f3bfd703fd487b33851537e0c067e6977231aa9c)

IJ-CR-156062

GitOrigin-RevId: c98c8c7f500423ee51d85d9490b7bae08ce7b522