IJPL-179075 Implement caching in TreeState for non-async trees
When performing a linear search for the matching child,
cache results so that they can be used for the next path.

This reduces the complexity of the search for nodes
with a lot of children from O(N^2) to O(N) because
on the next search, either the result is already cached,
or at least we skip the already-searched part,
starting from the next "unexplored" index instead.

GitOrigin-RevId: 7bbf2667649d959ba1d925c24cc2b4db3b5f5531