IJPL-44415 Update Linux custom frame button size on DFM toggle
Because the size of the header is not fully determined
by the Compact Mode alone, but also depends on the DFM,
just using CurrentTheme.TitlePane.buttonPreferredSize
isn't enough, as the DFM is not implemented through LaF
defaults.

We have to delegate to
CustomWindowHeaderUtil.getPreferredWindowHeaderHeight
here, which takes both modes into account.
We also need to add revalidate() / repaint() here,
as a DFM toggle event doesn't update the entire Swing
hierarchy, unlike the Compact Mode.
If we don't revalidate manually, the buttons won't update
until they're revalidated, e.g. due to a window move / resize.

GitOrigin-RevId: 91250887116c263b7f1fde42db60d55de3949c50